Complete Problem 4.9

+++ 04. Console-In-and-Out/01. Sum of 3 numbers/Program.cs
@@ -4,9 +4,9 @@ class Program
 {
     static void Main()
     {
-        double a = double.Parse(Console.ReadLine());
-        double b = double.Parse(Console.ReadLine());
-        double c = double.Parse(Console.ReadLine());
+        int a = int.Parse(Console.ReadLine());
+        int b = int.Parse(Console.ReadLine());
+        int c = int.Parse(Console.ReadLine());
         Console.WriteLine(a + b + c);
     }
 }
